implement trigger_run kill command via plugin system
