+++ python/michelangelo/cli/mactl/mactl.py
@@ -388,104 +388,11 @@ class CRD:
 
     def generate_kill(self, channel: Channel):
         """
-        Generate kill function that terminates resources by setting spec.kill=true.
-        Prompts for confirmation unless --yes flag is provided.
-
-        Examples: mactl trigger_run kill --namespace=ma-dev-test --name=training-pipeline-cron-trigger
-                 mactl pipeline_run kill --namespace=ma-dev-test --name=pipeline-run-20251016-210000-0523962d --yes
+        Generate kill function - delegated to plugins.
+        This is a placeholder that plugins will override.
         """
-        # Ensure get method is available for retrieving current resource
-        self.generate_get(channel)
-
         _LOG.info("Generate KILL method for %r / %r", self.name, self.full_name)
-        method_name, input_class, output_class = self._extract_method_info(
-            channel, self.full_name, "Update"
-        )
-
-        kill_func_signature = Signature(
-            [Parameter("self", Parameter.POSITIONAL_OR_KEYWORD)]
-            + [
-                Parameter(name, Parameter.POSITIONAL_OR_KEYWORD)
-                for name in ["namespace", "name"]
-            ]
-            + [Parameter("yes", Parameter.POSITIONAL_OR_KEYWORD, default=[])]
-        )
-
-        @bind_signature(kill_func_signature)
-        def kill_func(bound_args: Signature) -> Message:
-            _LOG.info("Start kill_func for %r", self.full_name)
-            _LOG.info("Bound arguments: %r", bound_args.arguments)
-            _self: CRD = bound_args.arguments["self"]
-            _name = get_single_arg(bound_args.arguments, "name")
-            _namespace = get_single_arg(bound_args.arguments, "namespace")
-            _yes = bound_args.arguments.get("yes", [])
-
-            # Confirmation logic
-            if not _yes:
-                resource_type = _self.name.replace("_", " ")
-                confirmation = input(f" > kill '{_name}' {resource_type}? [y/N] ")
-                if confirmation.lower() not in ['y', 'yes']:
-                    print("Kill operation cancelled.")
-                    return None
-
-            # Get current resource
-            current_resource = _self.get(_namespace, _name)
-            _LOG.info("Retrieved resource for kill: %r", current_resource)
-
-            # Create update request with kill flag set
-            current_dict = MessageToDict(
-                current_resource, preserving_proto_field_name=True
-            )
-
-            # Set kill flag in spec
-            resource_name = _self.name
-            if resource_name in current_dict and "spec" in current_dict[resource_name]:
-                current_dict[resource_name]["spec"]["kill"] = True
-            else:
-                _LOG.error("Missing required spec field in the resource structure")
-                raise ValueError(f"Cannot set kill flag on {resource_name}")
-
-            # Convert back to protobuf message for update
-            request_input = input_class()
-            ParseDict(current_dict, request_input, ignore_unknown_fields=True)
-
-            _LOG.info(
-                "KILL Request input (%r) ready: %r",
-                type(request_input),
-                request_input,
-            )
-
-            method_fullname = f"/{_self.full_name}/{method_name}"
-            _LOG.info("Method fullname for gRPC call: %s", method_fullname)
-
-            stub_method = channel.unary_unary(
-                method_fullname,
-                request_serializer=input_class.SerializeToString,
-                response_deserializer=output_class.FromString,
-            )
-
-            response = stub_method(
-                request_input,
-                metadata=METADATA_STUB,
-                timeout=30,
-            )
-
-            # Assert kill operation is successful by checking response spec
-            response_dict = MessageToDict(response, preserving_proto_field_name=True)
-            resource_name = _self.name
-            if (resource_name in response_dict and
-                "spec" in response_dict[resource_name] and
-                response_dict[resource_name]["spec"].get("kill") is True):
-                _LOG.info("Kill operation successfully set spec.kill=true")
-            else:
-                _LOG.error("Kill operation failed: spec.kill not set to true in response")
-                raise RuntimeError(f"Kill operation failed for {resource_name}: spec.kill not properly set")
-
-            _LOG.info("Kill operation completed (%r): %r", type(response), response)
-            return response
-
-        kill_func.__signature__ = kill_func_signature  # type: ignore[attr-defined]
-        self.kill = MethodType(kill_func, self)
+        pass
 
     def generate_list(self, channel: Channel):
         """
@@ -612,7 +519,7 @@ def get_message_class_by_name(pool: DescriptorPool, message_name: str) -> type[M
     return MessageClass
 
 
-def read_plugins(crd: CRD, user_command_action: str, crds: dict[str, CRD]) -> None:
+def read_plugins(crd: CRD, user_command_action: str, crds: dict[str, CRD], channel: Channel) -> None:
     """
     Read and apply plugins for a given crd.
     """
@@ -976,7 +883,7 @@ def main(channel: Channel):
         crds[user_command_crd], f"generate_{user_command_action}"
     )
     func_action_generator(channel)
-    read_plugins(crds[user_command_crd], user_command_action, crds)
+    read_plugins(crds[user_command_crd], user_command_action, crds, channel)
 
     assert hasattr(crds[user_command_crd], user_command_action)
     func_action = getattr(crds[user_command_crd], user_command_action)
